move tab row from the file to separate package
